perf: make an origin a byte range instead of two full positions
An `Origin` carried a line, a column, a byte offset, and a file kind for
each of its two ends: 40 bytes, of which 32 were derivable from the
other 8 and the source text. Every AST node, token, comment, control
directive, and error paid for it, and on a declaration-heavy file the
AST is over 90% of what a compile retains — 4000 typedefs in 230 KiB of
source parse to 104,008 nodes.

Origins are now `{ start, end, kind }` over byte offsets, and line and
column are recovered on demand through the new `LineIndex`, which is
what diagnostics and the language server already had the source text to
build. Nothing consulted stored line numbers for anything but display:
the one place that looked like it did, `LexerState::InsideControlDirective`,
never read its line payload, and `#line` records a directive without
remapping subsequent positions.

`NodeId` follows from `usize` to `u32` — several `NodeKind` variants
hold two or three, and that is what takes the widest variant from 72
bytes to 64.

  Node 128 -> 64, Token 56 -> 32, Comment 44 -> 20, Error 120 -> 72.

  230 KiB of typedefs : retained 17.4 MiB -> 7.1 MiB, peak 19.6 -> 10.2
  185 KiB, comment-heavy: retained 498 KiB -> 177 KiB, peak 1454 -> 269

Dropping the per-character line and column bookkeeping also makes the
lexer's `advance` a byte add, so this is faster as well: 767 us -> 620 us
on the comment-heavy input and 8.0 ms -> 5.8 ms on the typedefs.

Formatter output, diagnostics, and LSP hover ranges are byte-identical
to before on the test corpus. Two things did change, both fixes: the
`#!`-must-come-first error now names the file it is in rather than
"unknown", and a diagnostic from a fenced block in a markdown file now
carries the enclosing file's byte offsets and end line instead of
shifting only the start line and leaving the rest block-relative.

+++ src/type_checker.rs
@@ -423,7 +423,7 @@ mod tests {
                 origin: Origin::new_builtin(),
             },
         ];
-        let root = NodeId(nodes.len() - 1);
+        let root = NodeId(nodes.len() as u32 - 1);
 
         let mut node_to_type = HashMap::new();
         // Literal types are filled in before checking.
